pyproject.toml
[project]
name = "mein-projekt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pandas>=2.0.3",
    "python-dotenv>=1.0.1",
    "requests>=2.32.4",
    "supabase>=2.6.0",
]
